# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.IngestReadingBatchResponse:
    properties:
      accepted:
        type: integer
      rejected:
        type: integer
    type: object
...
  handlers.ingestReadingBatchRequest:
    properties:
      readings:
        items:
          $ref: '#/definitions/handlers.ingestReadingRequest'
        type: array
    type: object
  handlers.ingestReadingRequest:
    properties:
      device_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      metric_name:
        example: temperature
        type: string
      metric_value:
        example: 72.5
        type: number
      unit:
        example: celsius
        type: string
    type: object
...
  utils.Envelope:
    additionalProperties: {}
    type: object
host: localhost:8080
info:
  contact:
    email: support@strata.dev
    name: Strata Support
  description: ERP-CRM Hybrid API — direct competitor to Odoo.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Strata API
  version: 0.1.0
paths:
...
  /api/v1/iot/readings/batch:
    post:
      consumes:
      - application/json
      description: Processes multiple sensor readings from IoT devices in batch for
        high-frequency ingestion. Requires `iot.readings.ingest` permission.
      operationId: ingestReadingBatch
      parameters:
      - description: Batch IoT device readings payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.ingestReadingBatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.IngestReadingBatchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Envelope'
      security:
      - BearerAuth: []
      summary: Ingest device readings batch
      tags:
      - BI & IoT
...
securityDefinitions:
  ApiKeyAuth:
    description: API key for programmatic access (e.g. fleet telematics ingestion)
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'Enter your Bearer token in the format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...